Improve nuget release security (#16)
* Switch to OIDC-based NuGet authentication

Replaced static NuGet API key with a short-lived API key obtained via OpenID Connect (OIDC) for improved security. Updated the `Publish` step to use the dynamically generated API key and the `${{ env.NUGET_PATH }}` variable for `.nupkg` file paths. Added a new `NuGet login` step to handle OIDC authentication. Minor update to the `name` field in `Release.yml`.

* Update NuGet push command in Release.yml

Replaced the use of `${{ env.NUGET_PATH }}` with a wildcard pattern `**\*.nupkg` in the "Publish" step of `Release.yml`. This ensures that all `.nupkg` files in the directory and its subdirectories are pushed to the NuGet repository, improving flexibility and reliability.